Add klemol planner
This adds a new module called klemol_planner. It includes some
sort of an initial project structure. Transformations, placeholders
for different trajectory plannig algorithms, scripts, collision checker
and more are included.

It is made for testing purposes, not an actual demo, but gives a good
grasp how to implement it later